requirements.txt
spacy
streamlit
newspaper3k
pprint
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
lxml_html_clean